fix-file-structure(2nd commit)
